backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# AI APIs
anthropic==0.7.7
openai==1.3.7
google-generativeai==0.3.1

# AWS
boto3==1.29.7
botocore==1.32.7

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# ML/AI (for QC inspector)
numpy==1.26.2
tensorflow==2.15.0
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1

# CORS and middleware
# FastAPI has built-in CORS middleware, no additional package needed

# Validation
email-validator==2.1.0

# Caching (optional)
redis==5.0.1

# Logging
loguru==0.7.2

# Date/time
python-dateutil==2.8.2
